somali-early-warning-system/school_support_backend/requirements.txt
Django==6.0.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysql-connector-python
PyJWT
